fixed hexadecimal color code error
there was an error if uuser input a hexadecimal color code but without a '#', the app accepted the code but could not recognize it, just to show black color. Now fixed this error, users are required to add a '#' if the app recongize the input is a hexadecimal color code.